added tests

+++ database/migrations/2023_12_25_190920_delete_internet_access_role.php
@@ -13,7 +13,9 @@ return new class extends Migration {
      */
     public function up()
     {
-        DB::table('roles')->delete(Role::where('name', 'internet-user')->first()->id);
+        if (Role::where('name', 'internet-user')->first() !== null) {
+            DB::table('roles')->delete(Role::where('name', 'internet-user')->first()->id);
+        }
     }
 
     /**
